+++ mobile/src/components/profile/index.ts
@@ -5,3 +5,4 @@ export { FamilySection } from "./FamilySection";
 export { FeedbackSection } from "./FeedbackSection";
 export { ConfirmModal } from "./ConfirmModal";
 export { AiCreditsCard } from "./AiCreditsCard";
+export { StoreNamesSection } from "./StoreNamesSection";